+++ examples/03-dev/demo/lila.config.js
@@ -56,7 +56,12 @@ module.exports = {
     analyzerPort: 8191,
     localOptions: {
         senn: {
-            staticServerUrl: 'https//github.com/senntyou/lila'
+            staticServerUrl: 'https//github.com/senntyou/lila',
+            outResolveAlias: {
+                'bootstrap/dist/css/bootstrap.css': 'out_bootstrap/index.css',
+                bootstrap: 'out_bootstrap/index.js'
+
+            }
         }
     }
 };