+++ examples/03-dev/demo/project/src/test-5/index/index.js
@@ -1,4 +1,7 @@
 
+require('bootstrap/dist/css/bootstrap.css');
+require('bootstrap');
+
 require('base')();
 require('common')();