+++ project_config/handle/webpack.js
@@ -5,6 +5,7 @@ const _ = require('lodash');
 
 const staticServerUrl = require('../webpack/static_server_url');
 const resolveModules = require('../webpack/resolve_modules');
+const outResolveAlias = require('../webpack/out_resolve_alias');
 const makeDevConfig = require('../webpack/make_dev_config');
 const makeBuildConfig = require('../webpack/make_build_config');
 const makeAnalyzeConfig = require('../webpack/make_analyze_config');
@@ -15,6 +16,7 @@ module.exports = (config) => {
 
     staticServerUrl(config);
     resolveModules(config);
+    outResolveAlias(config);
 
     config.webpackDevConfig = makeDevConfig(config);
     config.webpackBuildConfig = makeBuildConfig(config);